Add new linter
The new linter is capable of detecting a lot more possible issues with
Moonscript compared to the current one. Similarly to the old linter, it
detects unused variables and undeclared global usages. It does however
detect unused declarations in a lot more contexts, such as import lists,
destructuring statements, for loops, parameter lists, etc.

The new linter also warns for shadowing declarations, i.e. a declaration
of a variable where a previous declaration already exists higher up in a
scope.